Move door into it's own file.
