+++ src/room/room.py
@@ -1,6 +1,5 @@
-from seika.assets import Texture
 from seika.math import Vector2
-from seika.node import Node2D, CollisionShape2D
+from seika.node import CollisionShape2D
 
 from src.project_properties import ProjectProperties
 
@@ -39,83 +38,6 @@ class WallColliders:
     def update_wall_positions(self, position: Vector2) -> None:
         for wall_collider in self.walls:
             wall_collider.position = position
-
-
-class DoorStatus:
-    OPEN = 0
-    CLOSED = 1
-
-
-class Door(CollisionShape2D):
-    Z_INDEX = 1
-    ROOM_LEFT_POSITION = Vector2(22, 74)
-    ROOM_RIGHT_POSITION = Vector2(334, 74)
-    ROOM_UP_POSITION = Vector2(168, 8)
-    ROOM_DOWN_POSITION = Vector2(168, 190)
-    OPEN_DOOR_TAG = ["open-door"]
-
-    def __init__(self, entity_id: int):
-        super().__init__(entity_id)
-        self.direction = Vector2()
-        self._is_open = False
-        self.sprite = None
-
-    def _get_dir_string(self) -> str:
-        if self.direction == Vector2.UP():
-            return "up"
-        elif self.direction == Vector2.DOWN():
-            return "down"
-        elif self.direction == Vector2.RIGHT():
-            return "right"
-        elif self.direction == Vector2.LEFT():
-            return "left"
-
-    def set_open(self, value: bool) -> None:
-        self._is_open = value
-        dir_string = self._get_dir_string()
-        if self._is_open:
-            is_open_string = "open"
-            self.tags = Door.OPEN_DOOR_TAG
-        else:
-            is_open_string = "closed"
-            self.tags = []
-        texture_file_path = (
-            f"assets/images/dungeon/door_{dir_string}_{is_open_string}.png"
-        )
-        self.sprite.texture = Texture.get(file_path=texture_file_path)
-
-    def is_open(self) -> bool:
-        self._is_open
-
-    @staticmethod
-    def new_door(dir: Vector2):
-        door = Door.new()
-        door.direction = dir
-        return door
-
-
-class DungeonDoors:
-    def __init__(
-        self,
-        left: Door,
-        right: Door,
-        up: Door,
-        down: Door,
-        container: Node2D,
-    ):
-        self.left = left
-        self.right = right
-        self.up = up
-        self.down = down
-        self.container = container
-        self.doors = [left, right, up, down]
-
-    def move(self, position: Vector2) -> None:
-        self.container.position = position
-
-    def update_tags(self, tags: list) -> None:
-        for door in self.doors:
-            door.tags = tags
 
 
 class Room:
