+++ src/room/room_builder.py
@@ -2,7 +2,8 @@ from seika.node import Node, Node2D, CollisionShape2D, Sprite
 from seika.math import Vector2, Rect2
 from seika.assets import Texture
 
-from src.room.room import WallColliders, DungeonDoors, Room, Door
+from src.room.room import Room, WallColliders
+from src.room.door import DungeonDoors, Door
 from src.room.room_manager import RoomManager
